requirements.txt
streamlit
pillow
requests
google-auth
rembg
onnxruntime

